broke out saving functionality into a child component
